Fixed image issues
