+++ src/components/Navbar.js
@@ -13,8 +13,8 @@ const Navbar = () => {
         <Nav>
             <Name>Alex Ocegueda Castro</Name>
             <NavLinks className='nav-links'>
-                <NavLink to="/about">About</NavLink>
-                <NavLink to="/">Projects</NavLink>
+                <NavLink to="/">About</NavLink>
+                <NavLink to="/projects">Projects</NavLink>
                 {/*<ResumeButton onClick={handleResumeClick}>Resume</ResumeButton>*/}
             </NavLinks>
         </Nav>
